feat: HL7Outbound is now Connection
- createConnection is now used
- clean up more of sendMessage
- fixed disconnect/reconnect issues (used live Corepoint) system for testing
- error class uses a string vs. a number now
- added 'autoConnect'
- pre lint issues

BREAKING CHANGE:

- HL7Outbound renamed Connection
