+++ __tests__/__utils__/index.ts
@@ -9,24 +9,3 @@ export const sleep = async (ms: number): Promise<unknown> => {
 export const expectEvent = async <T=any>(emitter: EventEmitter, name: string | symbol): Promise<T> => {
   return await new Promise<T>((resolve) => { emitter.once(name, resolve) })
 }
-
-/** @internal */
-export interface Deferred<T=any> {
-  resolve: (value: T | PromiseLike<T>) => void
-  reject: (reason?: any) => void
-  promise: Promise<T>
-}
-
-/** @internal */
-export const createDeferred = <T=any>(noUncaught?: boolean): Deferred<T> => {
-  const dfd: any = {}
-  dfd.promise = new Promise((resolve, reject) => {
-    dfd.resolve = resolve
-    dfd.reject = reject
-  })
-  /* istanbul ignore next */
-  if (noUncaught === false) {
-    dfd.promise.catch(() => {})
-  }
-  return dfd
-}
