+++ package.json
@@ -46,12 +46,12 @@
     "url": "git+https://github.com/Bugs5382/node-hl7-client.git"
   },
   "keywords": [
-    "HL7",
-    "HL7 Parser",
-    "HL7 Client",
-    "HL7 Builder",
-    "HL7 Speffications",
-    "HL7 Validation"
+    "hl7",
+    "hl7-parser",
+    "hl7-client",
+    "hl7-builder",
+    "hl7-speffications",
+    "hl7-validation"
   ],
   "author": "Shane Froebel",
   "license": "MIT",
@@ -64,30 +64,29 @@
     "@semantic-release/commit-analyzer": "^11.1.0",
     "@semantic-release/git": "^10.0.1",
     "@semantic-release/release-notes-generator": "^12.1.0",
-    "@the-rabbit-hole/semantic-release-config": "^1.4.0",
+    "@the-rabbit-hole/semantic-release-config": "^1.5.0",
     "@types/jest": "^29.5.11",
-    "@types/node": "^20.10.8",
+    "@types/node": "^20.11.10",
     "@types/tcp-port-used": "^1.0.4",
-    "@typescript-eslint/parser": "^6.18.1",
+    "@typescript-eslint/parser": "^6.20.0",
     "jest": "^29.7.0",
     "jest-ts-webcompat-resolver": "^1.0.0",
     "node-hl7-server": "^1.2.1",
-    "npm-check-updates": "^16.14.12",
+    "npm-check-updates": "^16.14.14",
     "npm-package-json-lint": "^7.1.0",
     "portfinder": "^1.0.32",
     "pre-commit": "^1.2.2",
-    "semantic-release": "^22.0.12",
+    "semantic-release": "^23.0.0",
     "snazzy": "^9.0.0",
     "tcp-port-used": "^1.0.2",
-    "ts-jest": "^29.1.1",
+    "ts-jest": "^29.1.2",
     "ts-node": "^10.9.2",
     "ts-standard": "^12.0.2",
-    "tsd": "^0.30.3",
+    "tsd": "^0.30.4",
     "typedoc": "^0.25.7",
     "typescript": "5.3.3"
   },
   "precommit": [
-    "lint:fix",
     "build"
   ]
 }
