+++ src/client/connection.ts
@@ -1,0 +1,442 @@
+import EventEmitter from 'node:events'
+import net, {Socket} from 'node:net'
+import {clearTimeout} from "node:timers";
+import tls from 'node:tls'
+import Batch from "../builder/batch.js";
+import FileBatch from "../builder/fileBatch.js";
+import Message from "../builder/message.js";
+import {PROTOCOL_MLLP_FOOTER, PROTOCOL_MLLP_HEADER} from "../utils/constants.js";
+import {ReadyState} from '../utils/enum.js'
+import {HL7Error, HL7FatalError} from '../utils/exception.js'
+import {ClientListenerOptions, normalizeClientListenerOptions, OutboundHandler} from '../utils/normalizedClient.js'
+import {createDeferred, Deferred, expBackoff} from '../utils/utils.js'
+import {Client} from './client.js'
+import {InboundResponse} from "./module/inboundResponse.js";
+
+export declare interface Connection {
+
+  /** The connection has been closed manually. You have to start the connection again. */
+  on(name: 'close', cb: () => void): this;
+  /** The connection is successfully (re)established or attempting to re-connect. */
+  on(name: 'connection', cb: () => void): this;
+  /** The connection has an error. */
+  on(name: 'error', cb: (err: any) => void): this;
+  /** The handle is open to do a manual start to connect. */
+  on(name: 'open', cb: (err: any) => void): this;
+
+
+  /** The total acknowledged for this connection. */
+  on(name: 'client.acknowledged', cb: (number: number) => void): this;
+  /** The total sent for this connection. */
+  on(name: 'client.sent', cb: (number: number) => void): this;
+}
+
+/** Connection Class
+ * @description Create a connection customer that will listen to result send to the particular port.
+ * @since 1.0.0 */
+export class Connection extends EventEmitter {
+  /** @internal */
+  _connectionTimer: NodeJS.Timeout | undefined
+  /** @internal */
+  _handler: OutboundHandler
+  /** @internal */
+  private readonly _main: Client
+  /** @internal */
+  private readonly _opt: ReturnType<typeof normalizeClientListenerOptions>
+  /** @internal */
+  private _retryCount: number
+  /** @internal */
+  _retryTimer: NodeJS.Timeout | undefined
+  /** @internal */
+  private _socket: Socket | undefined
+  /** @internal */
+  protected _readyState: ReadyState
+  /** @internal */
+  _pendingSetup: Promise<boolean> | boolean
+/*  /!** @internal *!/
+  private _responseBuffer: string */
+  /** @internal */
+  _onConnect: Deferred<void>
+  /** @internal */
+  private _awaitingResponse: boolean
+  /** @internal */
+  readonly stats = {
+    /** Total acknowledged messages back from server.
+     * @since 1.1.0 */
+    acknowledged: 0,
+    /** Total message sent to server.
+     * @since 1.1.0 */
+    sent: 0
+  }
+
+  /**
+   * @since 1.0.0
+   * @param client The client parent that we are connecting too.
+   * @param props The individual port connection options.
+   * Some values will be defaulted by the parent server connection.
+   * @param handler The function that will send the returned information back to the client after we got a response from the server.
+   * @example
+   * ```ts
+   * const OB = client.createOutbound({ port: 3000 }, async (res) => {})
+   * ```
+   */
+  constructor (client: Client, props: ClientListenerOptions, handler: OutboundHandler) {
+    super()
+
+    this._connectionTimer = undefined
+    this._handler = handler
+    this._main = client
+    this._awaitingResponse = false
+    /*this._nodeId = randomString(5)*/
+
+    this._opt = normalizeClientListenerOptions(props)
+
+    this._connect = this._connect.bind(this)
+
+    this._pendingSetup = true
+    this._retryCount = 0
+    this._retryTimer = undefined
+    this._onConnect = createDeferred(true)
+
+
+    if (this._opt.autoConnect) {
+      this._readyState = ReadyState.CONNECTING
+      this.emit('connecting')
+      this._socket = this._connect()
+    } else {
+      this._readyState = ReadyState.OPEN
+      this.emit('open')
+      this._socket = undefined
+    }
+
+
+  }
+
+  /** Close Client Listener Instance.
+   * @description Force close a connection.
+   * It Will stop any re-connection timers.
+   * If you want to restart, your app has to restart the connection.
+   * @since 1.0.0
+   * @example
+   * ```ts
+   * OB.close()
+   * ```
+   */
+  async close (): Promise<void> {
+
+    if (this._readyState === ReadyState.CLOSED) {
+      return // We are already closed. Nothing to do.
+    }
+
+    if (this._readyState === ReadyState.CLOSING) {
+      return new Promise(resolve => this._socket?.once('close', resolve))
+    }
+
+    if (this._readyState === ReadyState.CONNECTING) {
+      // mark that we set our internal that we are closing, so we do not try to re-connect
+      this._readyState = ReadyState.CLOSING
+      // clear retry timer
+      if (typeof this._retryTimer !== 'undefined') {
+        clearTimeout(this._retryTimer)
+      }
+      // let's clear out the try timer forcefully
+      this._retryTimer = undefined
+      // remove socket
+      this._socket?.destroy()
+      // return
+      return
+    }
+
+    // normal closing
+    this._readyState = ReadyState.CLOSING
+
+    this._socket?.end()
+
+    this.emit('close')
+
+    this._readyState = ReadyState.CLOSED
+
+  }
+
+  /**
+   * Start the connection if not auto started.
+   * @since 2.0.0
+   */
+  async start(): Promise<void> {
+
+    if (this._readyState === ReadyState.CONNECTING) {
+      return
+    }
+
+    if (this._readyState === ReadyState.CONNECTED) {
+      return
+    }
+
+    if (this._readyState === ReadyState.OPEN) {
+      return
+    }
+
+    if (this._readyState === ReadyState.CLOSING) {
+      return new Promise(resolve => this._socket?.once('close', resolve))
+    }
+
+    this.emit('connecting')
+
+    this._socket = this._connect()
+
+  }
+
+  /** Send a HL7 Message to the Listener
+   * @description This function sends a message/batch/file batch to the remote side.
+   * It has the ability, if set to auto-retry (defaulted to 1 re-connect before connection closes)
+   * @since 1.0.0
+   * @param message The message we need to send to the port.
+   * @example
+   * ```ts
+   *
+   * // the OB was set from the orginial 'createConnection' method.
+   *
+   * let message = new Message({
+   *  messageHeader: {
+   *    msh_9_1: "ADT",
+   *    msh_9_2: "A01",
+   *    msh_11_1: "P" // marked for production here in the example
+   *  }async sendMessage (message: Message | Batch | FileBatch): void {
+   * })
+   *
+   * await OB.sendMessage(message)
+   *
+   * ```
+   */
+  async sendMessage (message: Message | Batch | FileBatch): Promise<void> {
+
+    let attempts = 0
+    const maxAttempts = typeof this._opt.maxAttempts === 'undefined' ? this._main._opt.maxAttempts : this._opt.maxAttempts
+    const emitter = new EventEmitter()
+
+    const checkConnection = async (): Promise<boolean> => {
+      return this._readyState === ReadyState.CONNECTED
+    }
+
+    const checkAcknowledgement = async (): Promise<boolean> => {
+      return this._awaitingResponse
+    }
+
+    const checkSend = async (_message: string): Promise<boolean> => {
+      while (true) { // noinspection InfiniteLoopJS
+        try {
+          if ((this._readyState === ReadyState.CLOSED) || (this._readyState === ReadyState.CLOSING)) {
+            // noinspection ExceptionCaughtLocallyJS
+            throw new HL7FatalError('CONNECTION_INVALID_STATE', 'In an invalid state to be able to send message.')
+          }
+          if (this._readyState !== ReadyState.CONNECTED) {
+            // if we are not connected,
+            // check to see if we are now connected.
+            if (this._pendingSetup === false) {
+              this._pendingSetup = checkConnection().finally(() => {
+                this._pendingSetup = false
+              })
+            }
+          } else if (this._readyState === ReadyState.CONNECTED && this._opt.waitAck && this._awaitingResponse) {
+            // Ok, we ar now confirmed connected.
+            // However, since we are checking
+            // to make sure we wait for an ACKNOWLEDGEMENT from the server,
+            // that the message was gotten correctly from the last one we sent.
+            // We are still waiting, we need to recheck again
+            // if we are not connected,
+            // check to see if we are now connected.
+            if (this._pendingSetup === false) {
+              if (this._opt.waitAck) {
+                this._pendingSetup = checkAcknowledgement().finally(() => {
+                  this._pendingSetup = false
+                })
+              }
+            }
+          }
+          return await this._pendingSetup
+        } catch (err: any) {
+          Error.captureStackTrace(err)
+          if (++attempts >= maxAttempts) {
+            throw err
+          } else {
+            emitter.emit('retry', err)
+          }
+        }
+      }
+    }
+
+    // get the message
+    const theMessage = message.toString()
+
+    // check to see if we should be sending
+    await checkSend(theMessage)
+
+    // ok, if our options are to wait for an acknowledgement, set the var to "true"
+    if (this._opt.waitAck) {
+      this._awaitingResponse = true
+    }
+
+    // add MLLP settings to the message
+    const messageToSend = Buffer.from(`${PROTOCOL_MLLP_HEADER}${theMessage}${PROTOCOL_MLLP_FOOTER}`)
+
+    // send the message
+    this._socket?.write(messageToSend, this._opt.encoding, () => {
+      // we sent a message
+      ++this.stats.sent
+      // emit
+      this.emit('client.sent', this.stats.sent)
+    })
+
+  }
+
+  /**
+   * Read a file.
+   * @description We need to read a file.
+   * We are not doing anything else other than getting the {@link Buffer} of the file,
+   * so we can pass it onto the File Batch class to send it to the {@link sendMessage} method as a separate step
+   * @since 1.0.0
+   * @param fullFilePath The full file path of the file we need to read.
+   */
+/*  async readFile (fullFilePath: string): Promise<FileBatch> {
+    try {
+      const regex = /\n/mg
+      const subst = '\\r'
+      const fileBuffer = fs.readFileSync(fullFilePath)
+      const text = fileBuffer.toString().replace(regex, subst)
+      return new FileBatch({ text })
+    } catch (e: any) {
+      throw new HL7FatalError('', `Unable to read file: ${fullFilePath}`)
+    }
+  }*/
+
+  /** @internal */
+  private _connect  (): Socket {
+    let socket: Socket
+    const host = this._main._opt.host
+    const port = this._opt.port
+
+    this._retryTimer = undefined
+
+    if (typeof this._main._opt.tls !== 'undefined') {
+      socket = tls.connect({
+        host,
+        port,
+        ...this._main._opt.socket,
+        ...this._main._opt.tls
+      })
+    } else {
+      socket = net.connect({
+        host,
+        port
+      })
+    }
+
+    this._socket = socket
+
+    // set no delay
+    socket.setNoDelay(true)
+
+    let connectionError: Error | undefined
+
+    if (this._opt.connectionTimeout > 0) {
+      this._connectionTimer = setTimeout(() => {
+        socket.destroy(new HL7FatalError('CONNECTION_TIMEOUT', 'Connection timed out'))
+      }, this._opt.connectionTimeout)
+    }
+
+    socket.on('error', err => {
+      connectionError = connectionError || err
+    })
+
+    socket.on('close', () => {
+      if (this._readyState === ReadyState.CLOSING) {
+        this._readyState = ReadyState.CLOSED
+        this._reset()
+      } else {
+        connectionError = connectionError || new HL7Error('CONNECTION_CLOSE', 'Socket closed unexpectedly by server.')
+        if (this._readyState === ReadyState.OPEN) {
+          this._onConnect = createDeferred(true)
+        }
+        this._readyState = ReadyState.CONNECTING
+        this._reset()
+        const retryCount = this._retryCount++
+        const delay = expBackoff(this._opt.retryLow, this._opt.retryHigh, retryCount)
+        this._retryTimer = setTimeout(this._connect, delay)
+        if (retryCount <= 1) {
+          this.emit('error', connectionError)
+        }
+      }
+    })
+
+    socket.on('data', async (buffer) => {
+      // we got some sort of response, bad, good, or error,
+      // so lets tell the system we got "something"
+      this._awaitingResponse = false
+
+      socket.cork()
+
+      const indexOfVT = buffer.indexOf(PROTOCOL_MLLP_HEADER)
+      const indexOfFSCR = buffer.indexOf(PROTOCOL_MLLP_FOOTER)
+
+      let loadedMessage = buffer.toString().substring(indexOfVT, indexOfFSCR + 2)
+      loadedMessage = loadedMessage.replace(PROTOCOL_MLLP_HEADER, '')
+
+      if (loadedMessage.includes(PROTOCOL_MLLP_FOOTER)) {
+        // strip them out
+        loadedMessage = loadedMessage.replace(PROTOCOL_MLLP_FOOTER, '')
+        if (typeof this._handler !== 'undefined') {
+          // response
+          const response = new InboundResponse(loadedMessage)
+          // got an ACK, failure or not
+          ++this.stats.acknowledged
+          // update ack total
+          this.emit('client.acknowledged', this.stats.acknowledged)
+          // send it back
+          await this._handler(response)
+        }
+      }
+
+      socket.uncork()
+
+    })
+
+    const readerLoop = async () => {
+      try {
+        await this._negotiate()
+      } catch (err: any) {
+        if (err.code !== 'READ_END') {
+          socket.destroy(err)
+        }
+      }
+    }
+
+    readerLoop().then(_r => { /* noop */ })
+
+    return socket
+  }
+
+  /** @internal */
+  private async _negotiate (): Promise<void> {
+    if (this._socket?.writable) {
+      if (typeof this._connectionTimer !== 'undefined') {
+        clearTimeout(this._connectionTimer)
+      }
+      // we are open, not yet ready, but we can
+      this._readyState = ReadyState.OPEN
+      // let's go ahead and kill the connection timer
+      this._connectionTimer = undefined
+      // we are connected
+      this.emit('connection')
+    }
+  }
+
+  /** @internal */
+  private _reset (): void {
+    if (typeof this._connectionTimer !== 'undefined') {
+      clearTimeout(this._connectionTimer)
+    }
+    this._connectionTimer = undefined
+  }
+}
+
+export default Connection
