+++ src/index.ts
@@ -2,7 +2,7 @@ import { Client } from './client/client.js'
 import { Message } from './builder/message.js'
 import { Batch } from './builder/batch.js'
 import { FileBatch } from './builder/fileBatch.js'
-import { HL7Outbound } from './client/hl7Outbound.js'
+import { Connection } from './client/connection.js'
 import { Delimiters, ReadyState } from './utils/enum.js'
 import { OutboundHandler } from './utils/normalizedClient.js'
 import { InboundResponse } from './client/module/inboundResponse.js'
@@ -19,4 +19,4 @@ export type { ClientBuilderFileOptions, ClientBuilderMessageOptions, ClientBuild
 export type { HL7Error, HL7FatalError, HL7ParserError } from './utils/exception.js'
 
 export default Client
-export { Client, HL7Outbound, OutboundHandler, InboundResponse, FileBatch, Batch, Message, ReadyState, NodeBase, EmptyNode, Segment, Delimiters, HL7Node }
+export { Client, Connection, OutboundHandler, InboundResponse, FileBatch, Batch, Message, ReadyState, NodeBase, EmptyNode, Segment, Delimiters, HL7Node }
